fix: design style guide of the buttons
